Type error

+++ single_interface_rundown/Velocity_noise_1.4/S_and_T_fluxes.jl
@@ -85,7 +85,7 @@ function φ_molelcuar_flux!(flux_file::AbstractString, tracers::AbstractString, 
         Δz✶ = diff(z✶)
 
         φ_molecular_flux = Array{Float64}(undef, 3, length(timestamps))
-        interface_depth = Array{Int64}(undef, length(timestamps))
+        interface_depth = Array{Float64}(undef, length(timestamps))
 
         for i ∈ eachindex(timestamps)
